requirements.txt
openai
markdown
pygments
flask
pycryptodome
python-dotenv